final_project/requirements.txt
Flask==1.1.2
Flask-WTF==0.14.3
ibm-watson==6.1.0
python-dotenv==0.13.0